chore: move classes to the domain package
