+++ src/main/kotlin/lotto/domain/Lotties.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class Lotties(val value: List<Lotto>) {
 }
