+++ src/main/kotlin/lotto/domain/LottoGenerator.kt
@@ -1,4 +1,6 @@
-package lotto
+package lotto.domain
+
+import lotto.NumberGenerator
 
 
 class LottoGenerator(private val numberGenerator: NumberGenerator) {
